cad-generator/requirements.txt
Flask==2.0.3
requests
paho-mqtt
